split work on CLI into submodule and separate files
